+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added normalize_embeddings argument in sentence-transformers embedding models' call. 
 
-### Fixed
+### Removed
 - Removed request_id and uuid from Gemini & VertexAI chat formatters.
 
 ## [2.2.2] - 2025-07-15